+++ stock_monitor/data/stock/__init__.py
@@ -1,5 +1,5 @@
 """Stock data module"""
 
-from .stocks import format_stock_code, load_stock_data
+from .stocks import load_stock_data
 
-__all__ = ["load_stock_data", "format_stock_code"]
+__all__ = ["load_stock_data"]